// src/app/models/user.model.ts
export class User {
  id: number;
  username: '';
  first_name: '';
  last_name: '';
  email: '';
  password: string;
  mobile: string;
  position: string;
  photo: string;
  photo_thumb: string;
  bio: string;
  role: string;
  office: string;
  groups: any;
  offices: any;
  facebook: string;
  twitter: string;
  instagram: string;
  linkedin: string;
  website: string;
  crm_email: string;
  crm_password: string;
  permissions: any;
  level: any;
  added: Date;
}